moving file to the correct path
